SearchUI Reports - move classic report getter to its own class
